requirements.txt
yt-dlp
librosa
transformers
datasets
huggingface_hub
pydub
webvtt-py
torch
matplotlib
soundfile
audioread
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
ffmpeg-python
python-dotenv